Add SubiektGtId property to ProductCreateMinimalDto
A new property `SubiektGtId` of type `int?` has been added to the
`ProductCreateMinimalDto` class within the
`nopCommerceWebApiClient.Objects.Product` namespace. This property
is marked with the `[JsonIgnore]` attribute, ensuring it is excluded
from JSON serialization and deserialization. The `SubiektGtId` is
used internally to read the correct price and other data from
Subiekt GT, a business management software.